Add language and drawing support for activation boxes.
The syntax +x and -x represents activation and deactivation, respectively. Both
forms can be used wherever arcs can be used. The +x form starts an activation box
on the line of entity x, while -x ends it. Multiple activation boxes can be
stacked on top of each other.

There is a known issue with activation boxes and looping arcs (where the source
and destination entities are the same).